requirements.txt
numpy==1.23.5
openai_whisper==20230314
PyAudio==0.2.13
pygame==2.3.0
python-dotenv==1.0.0
requests==2.28.2
sounddevice==0.4.6
wavio==0.0.7
whisper==1.1.10